requirements.txt
aiofiles==0.8.0
anyio==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
coloredlogs==15.0.1
Deprecated==1.2.13
h11==0.12.0
httpcore==0.14.5
httpx==0.21.3
humanfriendly==10.0
idna==3.3
lxml==4.7.1
packaging==21.3
progressbar2==4.0.0
pyaria2==0.1.1
pyparsing==3.0.7
python-dotenv==0.19.2
python-utils==3.1.0
qbittorrent-api==2022.1.27
redis==4.1.1
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
urllib3==1.26.8
wrapt==1.13.3
yapf==0.32.0
